+++ practical_identifiability/src/pident/lib/__init__.py
@@ -1,3 +1,3 @@
-from pident.lib import models
+from pident.lib import liu_et_al_model, owens_bozic_model
 
-__all__ = ["models"]
+__all__ = ["liu_et_al_model", "owens_bozic_model"]
